Implement termination of streams after receiving a Goaway frame
The remote end indicates the highest stream id (initiated by the receiver of the
Goaway frame) which it processed. This means the receiver of the goaway frame knows
that all streams above that "last processed stream id" will be ignored by the server
and these requests can therefore be retried.

Currently we report a normal StreamError with a special sentence, that could be
made a special error in the future.

R=sgjesse@google.com

Review URL: https://chromereviews.googleplex.com/250127013 .